+++ packages/types/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.37] - 2026-07-22
+
+- Republish package following failed NPM publication
+
+## [1.0.36] - 2026-07-21
+
+- Republish package following failed NPM publication
+
+## [1.0.35] - 2026-06-16
+
+- Add Response types without and with file content
+
 ## [1.0.34] - 2026-06-11
 
 - Add lastEditedBy to the FormRecord type